core: move TestNativeGetMethod to the native package
